captcha_recognision/model/captcha_model.py
import tensorflow as tf
from keras.models import Sequential
from keras.layers import Activation, MaxPooling2D, Flatten, Conv2D, Dropout, Dense

from captcha_recognision.utils import console


class CaptchaModel:
    def __init__(self, model_path=None):
        self.model = self.load_model(model_path) if model_path else self.create_model()

    def create_model(self, input_shape: tuple, num_classes: int = 36):
        """
        Create a CNN model for CAPTCHA recognition.

        :param input_shape: Tuple specifying the input shape (height, width, channels)
        :param num_classes: Number of distinct characters in the CAPTCHA
        :return: CNN model
        """
        model = Sequential()

        # Convolutional layer 1
        model.add(Conv2D(32, (3, 3), activation="relu", input_shape=input_shape))
        model.add(Activation("relu"))
        model.add(MaxPooling2D(pool_size=(2, 2)))

        # Convolutional layer 2
        model.add(Conv2D(64, (3, 3), activation="relu"))
        model.add(Activation("relu"))
        model.add(MaxPooling2D((2, 2)))

        # Convolutional layer 3
        model.add(Conv2D(128, (3, 3), activation="relu"))
        model.add(Activation("relu"))
        model.add(MaxPooling2D((2, 2)))

        # Flattening and Dense layers
        model.add(Flatten())
        model.add(Dense(64, activation="relu"))
        model.add(Dropout(0.5))
        model.add(Dense(num_classes, activation="softmax"))

        # Compile the model
        model.compile(optimizer="adam", loss="categorical_crossentropy", metrics=["accuracy"])

        console.log(model.summary())

        return model

    def load_model(self, model_path):
        """
        Load a pre-trained model from the given path.

        :param model_path: Path to the model file
        :return: Loaded model
        """
        return tf.keras.models.load_model(model_path)

    def train(self, training_data, validation_data, epochs=10, batch_size=32):
        """
        Train the model on the given dataset.

        :param training_data: Training data
        :param validation_data: Validation data
        :param epochs: Number of epochs to train
        :param batch_size: Batch size for training
        """
        self.model.fit(training_data, epochs=epochs, batch_size=batch_size, validation_data=validation_data)

    def predict(self, character_images):
        """
        Predict characters from the segmented images.

        :param character_images: List of preprocessed character images
        :return: Predicted text
        """
        return "".join([self._predict_character(image) for image in character_images])

    def _predict_character(self, image):
        """
        Predict a single character from an image.

        :param image: Preprocessed character image
        :return: Predicted character
        """
        prediction = self.model.predict(image.reshape(1, *image.shape))
        return self._convert_prediction_to_char(prediction)

    def _convert_prediction_to_char(self, prediction):
        """
        Convert model prediction to character.

        :param prediction: Model prediction
        :return: Corresponding character
        """
        # Implement logic to convert prediction to character
        pass
